+++ dvc/data/reference.py
@@ -8,10 +8,9 @@ from dvc.objects.errors import ObjectFormatError
 from dvc.objects.file import HashFile
 
 if TYPE_CHECKING:
-    from dvc.fs.base import FileSystem
+    from dvc.fs.base import AnyFSPath, FileSystem
     from dvc.hash_info import HashInfo
     from dvc.objects.db import ObjectDB
-    from dvc.types import AnyPath
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +23,7 @@ class ReferenceHashFile(HashFile):
 
     def __init__(
         self,
-        fs_path: "AnyPath",
+        fs_path: "AnyFSPath",
         fs: "FileSystem",
         hash_info: "HashInfo",
         checksum: Optional[str] = None,
