+++ dvc/fs/base.py
@@ -1,10 +1,22 @@
 import contextlib
 import logging
 import os
+import shutil
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from functools import partialmethod
 from multiprocessing import cpu_count
-from typing import ClassVar, Dict, Optional
+from typing import (
+    IO,
+    TYPE_CHECKING,
+    Any,
+    ClassVar,
+    Dict,
+    Iterator,
+    List,
+    Optional,
+    Union,
+    overload,
+)
 
 from funcy import cached_property
 from tqdm.utils import CallbackIOWrapper
@@ -15,7 +27,20 @@ from dvc.ui import ui
 from dvc.utils import tmp_fname
 from dvc.utils.fs import makedirs, move
 
+if TYPE_CHECKING:
+    from fsspec.spec import AbstractFileSystem
+    from typing_extensions import Literal
+
+
 logger = logging.getLogger(__name__)
+
+
+FSPath = str
+AnyFSPath = str
+
+# An info() entry, might evolve to a TypedDict
+# in the future (e.g for properly type 'size' etc).
+Entry = Dict[str, Any]
 
 
 class RemoteActionNotImplemented(DvcException):
@@ -38,7 +63,7 @@ class FileSystem:
     HASH_JOBS = max(1, min(4, cpu_count() // 2))
     LIST_OBJECT_PAGE_SIZE = 1000
     TRAVERSE_WEIGHT_MULTIPLIER = 5
-    TRAVERSE_PREFIX_LEN = 3
+    TRAVERSE_PREFIX_LEN = 2
     TRAVERSE_THRESHOLD_SIZE = 500000
     CAN_TRAVERSE = True
 
@@ -53,9 +78,11 @@ class FileSystem:
         self.jobs = kwargs.get("jobs") or self._JOBS
         self.hash_jobs = kwargs.get("checksum_jobs") or self.HASH_JOBS
         self._config = kwargs
+        self.fs_args = {"skip_instance_cache": True}
+        self.fs_args.update(self._prepare_credentials(**kwargs))
 
     @property
-    def config(self):
+    def config(self) -> Dict[str, Any]:
         return self._config
 
     @cached_property
@@ -65,21 +92,37 @@ class FileSystem:
         return Path(self.sep)
 
     @classmethod
-    def _strip_protocol(cls, path: str):
+    def _strip_protocol(cls, path: str) -> str:
         return path
 
-    def unstrip_protocol(self, path):
+    def unstrip_protocol(self, path: str) -> str:
         return path
+
+    @cached_property
+    def fs(self) -> "AbstractFileSystem":
+        raise NotImplementedError
 
     @staticmethod
-    def _get_kwargs_from_urls(urlpath):  # pylint:disable=unused-argument
+    def _get_kwargs_from_urls(urlpath: str) -> "Dict[str, Any]":
+        from fsspec.utils import infer_storage_options
+
+        options = infer_storage_options(urlpath)
+        options.pop("path", None)
+        options.pop("protocol", None)
+        return options
+
+    def _prepare_credentials(
+        self, **config: Dict[str, Any]  # pylint: disable=unused-argument
+    ) -> Dict[str, Any]:
+        """Prepare the arguments for authentication to the
+        host filesystem"""
         return {}
 
     @classmethod
-    def get_missing_deps(cls):
+    def get_missing_deps(cls) -> List[str]:
         import importlib
 
-        missing = []
+        missing: List[str] = []
         for package, module in cls.REQUIRES.items():
             try:
                 importlib.import_module(module)
@@ -127,90 +170,155 @@ class FileSystem:
             f"dependencies: {missing}. {hint}"
         )
 
-    def checksum(self, path) -> str:
-        raise NotImplementedError
+    def isdir(self, path: AnyFSPath) -> bool:
+        return self.fs.isdir(path)
 
-    def open(self, path, mode: str = "r", encoding: str = None, **kwargs):
-        raise RemoteActionNotImplemented("open", self.scheme)
+    def isfile(self, path: AnyFSPath) -> bool:
+        return self.fs.isfile(path)
 
-    def exists(self, path) -> bool:
-        raise NotImplementedError
+    def is_empty(self, path: AnyFSPath) -> bool:
+        entry = self.info(path)
+        if entry["type"] == "directory":
+            return not self.fs.ls(path)
+        return entry["size"] == 0
 
-    # pylint: disable=unused-argument
+    def open(
+        self,
+        path: AnyFSPath,
+        mode: str = "r",
+        encoding: Optional[str] = None,
+        **kwargs,
+    ) -> "IO":  # pylint: disable=arguments-differ
+        return self.fs.open(path, mode=mode, encoding=encoding, **kwargs)
 
-    def isdir(self, path):
-        """Optional: Overwrite only if the remote has a way to distinguish
-        between a directory and a file.
-        """
-        return False
+    def checksum(self, path: AnyFSPath) -> str:
+        return self.fs.checksum(path)
 
-    def isfile(self, path):
-        """Optional: Overwrite only if the remote has a way to distinguish
-        between a directory and a file.
-        """
-        return True
+    def copy(self, from_info: AnyFSPath, to_info: AnyFSPath) -> None:
+        self.makedirs(self.path.parent(to_info))
+        self.fs.copy(from_info, to_info)
 
-    def iscopy(self, path):
-        """Check if this file is an independent copy."""
-        return False  # We can't be sure by default
+    def exists(self, path: AnyFSPath) -> bool:
+        return self.fs.exists(path)
 
-    def walk(self, top, topdown=True, **kwargs):
-        """Return a generator with (root, dirs, files)."""
-        raise NotImplementedError
+    def lexists(self, path: AnyFSPath) -> bool:
+        return self.fs.lexists(path)
 
-    def find(self, path, prefix=None):
-        """Return a generator with `str`s to all the files.
+    def symlink(self, from_info: AnyFSPath, to_info: AnyFSPath) -> None:
+        try:
+            return self.fs.symlink(from_info, to_info)
+        except AttributeError:
+            raise RemoteActionNotImplemented("symlink", self.scheme)
 
-        Optional kwargs:
-            prefix (bool): If true `path` will be treated as a prefix
-                rather than directory path.
-        """
-        raise NotImplementedError
+    def hardlink(self, from_info: AnyFSPath, to_info: AnyFSPath) -> None:
+        try:
+            return self.fs.hardlink(from_info, to_info)
+        except AttributeError:
+            raise RemoteActionNotImplemented("hardlink", self.scheme)
 
-    def is_empty(self, path):
-        return False
+    def reflink(self, from_info: AnyFSPath, to_info: AnyFSPath) -> None:
+        try:
+            return self.fs.reflink(from_info, to_info)
+        except AttributeError:
+            raise RemoteActionNotImplemented("reflink", self.scheme)
 
-    def info(self, path):
-        raise NotImplementedError
+    def is_symlink(self, path: AnyFSPath) -> bool:
+        try:
+            return self.fs.is_symlink(path)
+        except AttributeError:
+            return False
 
-    def getsize(self, path):
+    def is_hardlink(self, path: AnyFSPath) -> bool:
+        try:
+            return self.fs.is_hardlink(path)
+        except AttributeError:
+            return False
+
+    def iscopy(self, path: AnyFSPath) -> bool:
+        return self.is_symlink(path) or self.is_hardlink(path)
+
+    @overload
+    def ls(
+        self, path: AnyFSPath, detail: "Literal[True]"
+    ) -> "Iterator[Entry]":
+        ...
+
+    @overload
+    def ls(self, path: AnyFSPath, detail: "Literal[False]") -> Iterator[str]:
+        ...
+
+    def ls(self, path, detail=False, **kwargs):
+        return self.fs.ls(path, detail=detail)
+
+    def find(
+        self,
+        path: AnyFSPath,
+        prefix: bool = False,  # pylint: disable=unused-argument
+    ) -> Iterator[str]:
+        yield from self.fs.find(path)
+
+    def move(self, from_info: AnyFSPath, to_info: AnyFSPath) -> None:
+        self.fs.move(from_info, to_info)
+
+    def remove(self, path: AnyFSPath) -> None:
+        self.fs.rm_file(path)
+
+    def info(self, path: AnyFSPath) -> "Entry":
+        return self.fs.info(path)
+
+    def makedirs(self, path: AnyFSPath, **kwargs: Any) -> None:
+        self.fs.makedirs(path, exist_ok=kwargs.pop("exist_ok", True))
+
+    def put_file(
+        self,
+        from_file: AnyFSPath,
+        to_info: AnyFSPath,
+        callback: Any = DEFAULT_CALLBACK,
+        **kwargs,
+    ) -> None:
+        self.fs.put_file(from_file, to_info, callback=callback, **kwargs)
+        self.fs.invalidate_cache(self.path.parent(to_info))
+
+    def get_file(
+        self,
+        from_info: AnyFSPath,
+        to_info: AnyFSPath,
+        callback: Any = DEFAULT_CALLBACK,
+        **kwargs,
+    ) -> None:
+        self.fs.get_file(from_info, to_info, callback=callback, **kwargs)
+
+    def upload_fobj(self, fobj: IO, to_info: AnyFSPath, **kwargs) -> None:
+        self.makedirs(self.path.parent(to_info))
+        with self.open(to_info, "wb") as fdest:
+            shutil.copyfileobj(
+                fobj,
+                fdest,
+                length=getattr(fdest, "blocksize", None),  # type: ignore
+            )
+
+    def walk(
+        self,
+        path: AnyFSPath,
+        topdown: bool = True,
+        **kwargs: Any,
+    ):
+        return self.fs.walk(path, topdown=topdown, **kwargs)
+
+    def getsize(self, path: AnyFSPath) -> Optional[int]:
         return self.info(path).get("size")
-
-    def remove(self, path):
-        raise RemoteActionNotImplemented("remove", self.scheme)
-
-    def makedirs(self, path, **kwargs):
-        """Optional: Implement only if the remote needs to create
-        directories before copying/linking/moving data
-        """
-
-    def move(self, from_info, to_info):
-        self.copy(from_info, to_info)
-        self.remove(from_info)
-
-    def copy(self, from_info, to_info):
-        raise RemoteActionNotImplemented("copy", self.scheme)
-
-    def symlink(self, from_info, to_info):
-        raise RemoteActionNotImplemented("symlink", self.scheme)
-
-    def hardlink(self, from_info, to_info):
-        raise RemoteActionNotImplemented("hardlink", self.scheme)
-
-    def reflink(self, from_info, to_info):
-        raise RemoteActionNotImplemented("reflink", self.scheme)
 
     # pylint: enable=unused-argument
 
     def upload(
         self,
-        from_info,
-        to_info,
-        total=None,
-        desc=None,
+        from_info: Union[AnyFSPath, IO],
+        to_info: AnyFSPath,
+        total: int = None,
+        desc: str = None,
         callback=None,
-        no_progress_bar=False,
-        **pbar_args,
+        no_progress_bar: bool = False,
+        **pbar_args: Any,
     ):
         is_file_obj = hasattr(from_info, "read")
         method = "upload_fobj" if is_file_obj else "put_file"
@@ -246,6 +354,7 @@ class FileSystem:
                 # pylint: disable=no-member
                 return self.upload_fobj(wrapped, to_info, size=total)
 
+            assert isinstance(from_info, str)
             logger.debug("Uploading '%s' to '%s'", from_info, to_info)
             # pylint: disable=no-member
             return self.put_file(
@@ -254,14 +363,14 @@ class FileSystem:
 
     def download(
         self,
-        from_info,
-        to_info,
-        name=None,
+        from_info: AnyFSPath,
+        to_info: AnyFSPath,
+        name: str = None,
         callback=None,
-        no_progress_bar=False,
-        jobs=None,
-        _only_file=False,
-        **kwargs,
+        no_progress_bar: bool = False,
+        jobs: int = None,
+        _only_file: bool = False,
+        **kwargs: Any,
     ):
         from .local import localfs
 
@@ -295,10 +404,10 @@ class FileSystem:
 
     def _download_dir(
         self,
-        from_info,
-        to_info,
+        from_info: AnyFSPath,
+        to_info: AnyFSPath,
         callback=DEFAULT_CALLBACK,
-        jobs=None,
+        jobs: int = None,
         **kwargs,
     ):
         from .local import localfs
@@ -339,10 +448,10 @@ class FileSystem:
 
     def _download_file(
         self,
-        from_info,
-        to_info,
+        from_info: AnyFSPath,
+        to_info: AnyFSPath,
         callback=DEFAULT_CALLBACK,
-    ):
+    ) -> None:
         from .local import localfs
 
         makedirs(localfs.path.parent(to_info), exist_ok=True)
@@ -360,3 +469,59 @@ class FileSystem:
             raise
 
         move(tmp_file, to_info)
+
+
+class ObjectFileSystem(FileSystem):  # pylint: disable=abstract-method
+    TRAVERSE_PREFIX_LEN = 3
+
+    def makedirs(self, path: AnyFSPath, **kwargs: Any) -> None:
+        # For object storages make this method a no-op. The original
+        # fs.makedirs() method will only check if the bucket exists
+        # and create if it doesn't though we don't want to support
+        # that behavior, and the check will cost some time so we'll
+        # simply ignore all mkdir()/makedirs() calls.
+        return None
+
+    def _isdir(self, path: AnyFSPath) -> bool:
+        # Directory in object storages are interpreted differently
+        # among different fsspec providers, so this logic is a temporary
+        # measure for us to adapt as of now. It checks whether it is a
+        # directory (as in a prefix with contents) or whether it is an empty
+        # file where it's name ends with a forward slash
+
+        entry = self.info(path)
+        return entry["type"] == "directory" or (
+            entry["size"] == 0
+            and entry["type"] == "file"
+            and entry["name"].endswith("/")
+        )
+
+    def isdir(self, path: AnyFSPath) -> bool:
+        try:
+            return self._isdir(path)
+        except FileNotFoundError:
+            return False
+
+    def isfile(self, path: AnyFSPath) -> bool:
+        try:
+            return not self._isdir(path)
+        except FileNotFoundError:
+            return False
+
+    def find(self, path: AnyFSPath, prefix: bool = False) -> Iterator[str]:
+        if prefix:
+            with_prefix = self.path.parent(path)
+            files = self.fs.find(with_prefix, prefix=self.path.parts(path)[-1])
+        else:
+            with_prefix = path
+            files = self.fs.find(path)
+
+        # When calling find() on a file, it returns the same file in a list.
+        # For object-based storages, the same behavior applies to empty
+        # directories since they are represented as files. This condition
+        # checks whether we should yield an empty list (if it is an empty
+        # directory) or just yield the file itself.
+        if len(files) == 1 and files[0] == with_prefix and self.isdir(path):
+            return None
+
+        yield from files
