+++ dvc/fs/s3.py
@@ -8,13 +8,13 @@ from funcy import cached_property, wrap_prop
 from dvc.scheme import Schemes
 
 from ._callback import DEFAULT_CALLBACK
-from .fsspec_wrapper import ObjectFSWrapper
+from .base import ObjectFileSystem
 
 _AWS_CONFIG_PATH = os.path.join(os.path.expanduser("~"), ".aws", "config")
 
 
 # pylint:disable=abstract-method
-class BaseS3FileSystem(ObjectFSWrapper):
+class BaseS3FileSystem(ObjectFileSystem):
     scheme = Schemes.S3
     REQUIRES = {"s3fs": "s3fs", "boto3": "boto3"}
     PARAM_CHECKSUM = "etag"
