+++ dvc/fs/webhdfs.py
@@ -5,10 +5,10 @@ from funcy import cached_property, wrap_prop
 from dvc.scheme import Schemes
 
 # pylint:disable=abstract-method
-from .fsspec_wrapper import FSSpecWrapper
+from .base import FileSystem
 
 
-class WebHDFSFileSystem(FSSpecWrapper):
+class WebHDFSFileSystem(FileSystem):
     scheme = Schemes.WEBHDFS
     REQUIRES = {"fsspec": "fsspec"}
     PARAM_CHECKSUM = "checksum"
